Revert to commit 507d0ac

+++ include/operations/file_operations.h
@@ -13,37 +13,37 @@
 
 /* Add entry to file list */
 int add_entry_to_file_list(const char *input_file, const char *output_file,
-                           unsigned long long original_size,
-                           unsigned long long processed_size,
+                           unsigned long original_size, 
+                           unsigned long processed_size,
                            int quiet);
 
 /* Encrypt file using ChaCha20 */
-unsigned long long encrypt_file(const char *input_file, const char *output_file,
-                                const char *password, int quiet,
-                                unsigned long long *original_size_out);
+unsigned long encrypt_file(const char *input_file, const char *output_file,
+                           const char *password, int quiet,
+                           unsigned long *original_size_out);
 
 /* Decrypt file using ChaCha20 */
-unsigned long long decrypt_file(const char *input_file, const char *output_file,
-                                const char *password, int quiet,
-                                unsigned long long *original_size_out);
+unsigned long decrypt_file(const char *input_file, const char *output_file,
+                           const char *password, int quiet,
+                           unsigned long *original_size_out);
 
 /* Compress file using Huffman coding */
-unsigned long long compress_file(const char *input_file, const char *output_file,
-                                 int quiet, unsigned long long *original_size_out);
+unsigned long compress_file(const char *input_file, const char *output_file,
+                            int quiet, unsigned long *original_size_out);
 
 /* Decompress file using Huffman coding */
-unsigned long long decompress_file(const char *input_file, const char *output_file,
-                                   int quiet, unsigned long long *original_size_out);
+unsigned long decompress_file(const char *input_file, const char *output_file,
+                              int quiet, unsigned long *original_size_out);
 
 /* Process file (compress then encrypt) */
-unsigned long long process_file(const char *input_file, const char *output_file,
-                                const char *password, int quiet,
-                                unsigned long long *original_size_out);
+unsigned long process_file(const char *input_file, const char *output_file,
+                           const char *password, int quiet,
+                           unsigned long *original_size_out);
 
 /* Extract file (decrypt then decompress) */
-unsigned long long extract_file(const char *input_file, const char *output_file,
-                                const char *password, int quiet,
-                                unsigned long long *original_size_out);
+unsigned long extract_file(const char *input_file, const char *output_file,
+                           const char *password, int quiet,
+                           unsigned long *original_size_out);
 
 /* Handle file list operations */
 int handle_file_list(const char *command, const char *filename_pattern,
